Renamed internal directory to lib; Upgrade Golang version to 1.18; Update all dependencies; Update golangci-lint config; Update README
