+++ .golangci.yml
@@ -1,26 +1,22 @@
+---
 # golangci-lint configuration
 #
 linters:
   disable-all: true
   enable:
-    - deadcode
     - errcheck
     - gosimple
     - govet
     - ineffassign
     - staticcheck
-    - structcheck
     - typecheck
     - unused
-    - varcheck
     - bodyclose
     - dogsled
     - gomnd
     - gosec
-    - maligned
     - misspell
     - rowserrcheck
-    - scopelint
     - unconvert
     - unparam
     - whitespace
@@ -28,12 +24,12 @@ linters-settings:
   gomnd:
     settings:
       mnd:
-        checks: argument,case,condition,return
+        checks:
+          - argument
+          - case
+          - condition
+          - return
   govet:
     check-shadowing: true
-  maligned:
-    suggest-new: true
   unparam:
     check-exported: true
-  unused:
-    check-exported: false
