+++ Dockerfile
@@ -1,7 +1,7 @@
 ARG APPVERSION
 
-FROM golang:1.13-buster AS builder
-RUN apt-get update && apt-get install --yes upx-ucl
+FROM golang:1.18 AS builder
+RUN apt-get update --yes && apt-get install --yes upx-ucl
 WORKDIR /tmp/src/fdwctl
 COPY . .
 RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -ldflags "-s -w -X github.com/neflyte/fdwctl/cmd/fdwctl/cmd.AppVersion=${APPVERSION}" -o fdwctl ./cmd/fdwctl
