+++ README.md
@@ -2,8 +2,8 @@
 A PostgreSQL Foreign Data Wrapper (FDW) management CLI
 -
 
-![Golang v1.14+](https://img.shields.io/badge/Golang-v1.14+-blue?style=for-the-badge&logo=go&color=00add8&link=https://golang.org)
-![PostgreSQL v10+](https://img.shields.io/badge/PostgreSQL-v10+-blue?style=for-the-badge&logo=postgresql&color=336791&link=https://postgresql.org)
+![Golang v1.18+](https://img.shields.io/badge/Golang-v1.18+-blue?style=for-the-badge&logo=go&color=00add8&link=https://go.dev)
+![PostgreSQL v15+](https://img.shields.io/badge/PostgreSQL-v15+-blue?style=for-the-badge&logo=postgresql&color=336791&link=https://postgresql.org)
 
 [Features](#features) | [Installation](#installation) | [Usage](#usage) | [Configuration](#configuration)
