+++ cmd/fdwctl/cmd/desiredstate.go
@@ -4,11 +4,11 @@ import (
 	"context"
 	"database/sql"
 
-	"github.com/neflyte/fdwctl/internal/config"
-	"github.com/neflyte/fdwctl/internal/database"
-	"github.com/neflyte/fdwctl/internal/logger"
-	"github.com/neflyte/fdwctl/internal/model"
-	"github.com/neflyte/fdwctl/internal/util"
+	"github.com/neflyte/fdwctl/lib/config"
+	"github.com/neflyte/fdwctl/lib/database"
+	"github.com/neflyte/fdwctl/lib/logger"
+	"github.com/neflyte/fdwctl/lib/model"
+	"github.com/neflyte/fdwctl/lib/util"
 	"github.com/spf13/cobra"
 )
