+++ lib/logger/elastic.go
